Split helper classes into files
