+++ USE_CASES.md
@@ -20,19 +20,19 @@ require 'vendor/autoload.php';
 // If you are not using Composer
 // require("path/to/sendgrid-php/sendgrid-php.php");
 
-$from = new SendGrid\Email("Example User", "test@example.com");
+$from = new \Helpers\Email("Example User", "test@example.com");
 $subject = "Sending with SendGrid is Fun";
-$to = new SendGrid\Email("Example User", "test@example.com");
-$content = new SendGrid\Content("text/plain", "and easy to do anywhere, even with PHP");
+$to = new \Helpers\Email("Example User", "test@example.com");
+$content = new \Helpers\Content("text/plain", "and easy to do anywhere, even with PHP");
 $file = 'my_file.txt';
 $file_encoded = base64_encode(file_get_contents($file));
-$attachment = new SendGrid\Attachment();
+$attachment = new \Helpers\Attachment();
 $attachment->setContent($file_encoded);
 $attachment->setType("application/text");
 $attachment->setDisposition("attachment");
 $attachment->setFilename("my_file.txt");
 
-$mail = new SendGrid\Mail($from, $subject, $to, $content);
+$mail = new \Helpers\Mail($from, $subject, $to, $content);
 $mail->addAttachment($attachment);
 
 $apiKey = getenv('SENDGRID_API_KEY');
@@ -93,11 +93,11 @@ require 'vendor/autoload.php';
 // If you are not using Composer (recommended)
 // require("path/to/sendgrid-php/sendgrid-php.php");
 
-$from = new SendGrid\Email(null, "test@example.com");
+$from = new \Helpers\Email(null, "test@example.com");
 $subject = "I'm replacing the subject tag";
-$to = new SendGrid\Email(null, "test@example.com");
-$content = new SendGrid\Content("text/html", "I'm replacing the <strong>body tag</strong>");
-$mail = new SendGrid\Mail($from, $subject, $to, $content);
+$to = new \Helpers\Email(null, "test@example.com");
+$content = new \Helpers\Content("text/html", "I'm replacing the <strong>body tag</strong>");
+$mail = new \Helpers\Mail($from, $subject, $to, $content);
 $mail->personalization[0]->addSubstitution("-name-", "Example User");
 $mail->personalization[0]->addSubstitution("-city-", "Denver");
 $mail->setTemplateId("13b8f94f-bcae-4ec6-b752-70d6cb59f932");
