+++ lib/Helpers/ClickTracking.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * This helper builds the request body for a /mail/send API call.
+ *
+ * PHP version 5.6, 7
+ *
+ * @author    Elmer Thomas <dx@sendgrid.com>
+ * @copyright 2017 SendGrid
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @version   GIT: <git_id>
+ * @link      http://packagist.org/packages/sendgrid/sendgrid
+ */
+
+namespace Helpers;
+
+class ClickTracking implements \JsonSerializable
+{
+    private $enable;
+    private $enable_text;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setEnableText($enable_text)
+    {
+        $this->enable_text = $enable_text;
+    }
+
+    public function getEnableText()
+    {
+        return $this->enable_text;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable'      => $this->getEnable(),
+                'enable_text' => $this->getEnableText()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
