+++ lib/Helpers/Footer.php
@@ -1,0 +1,65 @@
+<?php
+/**
+ * This helper builds the request body for a /mail/send API call.
+ *
+ * PHP version 5.6, 7
+ *
+ * @author    Elmer Thomas <dx@sendgrid.com>
+ * @copyright 2017 SendGrid
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @version   GIT: <git_id>
+ * @link      http://packagist.org/packages/sendgrid/sendgrid
+ */
+
+namespace Helpers;
+
+class Footer implements \JsonSerializable
+{
+    private $enable;
+    private $text;
+    private $html;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setText($text)
+    {
+        $this->text = $text;
+    }
+
+    public function getText()
+    {
+        return $this->text;
+    }
+
+    public function setHtml($html)
+    {
+        $this->html = $html;
+    }
+
+    public function getHtml()
+    {
+        return $this->html;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable' => $this->getEnable(),
+                'text'   => $this->getText(),
+                'html'   => $this->getHtml()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
