+++ lib/Helpers/MailSettings.php
@@ -1,0 +1,89 @@
+<?php
+/**
+ * This helper builds the request body for a /mail/send API call.
+ *
+ * PHP version 5.6, 7
+ *
+ * @author    Elmer Thomas <dx@sendgrid.com>
+ * @copyright 2017 SendGrid
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @version   GIT: <git_id>
+ * @link      http://packagist.org/packages/sendgrid/sendgrid
+ */
+
+namespace Helpers;
+
+class MailSettings implements \JsonSerializable
+{
+    private $bcc;
+    private $bypass_list_management;
+    private $footer;
+    private $sandbox_mode;
+    private $spam_check;
+
+    public function setBccSettings($bcc)
+    {
+        $this->bcc = $bcc;
+    }
+
+    public function getBccSettings()
+    {
+        return $this->bcc;
+    }
+
+    public function setBypassListManagement($bypass_list_management)
+    {
+        $this->bypass_list_management = $bypass_list_management;
+    }
+
+    public function getBypassListManagement()
+    {
+        return $this->bypass_list_management;
+    }
+
+    public function setFooter($footer)
+    {
+        $this->footer = $footer;
+    }
+
+    public function getFooter()
+    {
+        return $this->footer;
+    }
+
+    public function setSandboxMode($sandbox_mode)
+    {
+        $this->sandbox_mode = $sandbox_mode;
+    }
+
+    public function getSandboxMode()
+    {
+        return $this->sandbox_mode;
+    }
+
+    public function setSpamCheck($spam_check)
+    {
+        $this->spam_check = $spam_check;
+    }
+
+    public function getSpamCheck()
+    {
+        return $this->spam_check;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'bcc'                    => $this->getBccSettings(),
+                'bypass_list_management' => $this->getBypassListManagement(),
+                'footer'                 => $this->getFooter(),
+                'sandbox_mode'           => $this->getSandboxMode(),
+                'spam_check'             => $this->getSpamCheck()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
