+++ lib/Helpers/OpenTracking.php
@@ -1,0 +1,53 @@
+<?php
+/**
+ * This helper builds the request body for a /mail/send API call.
+ *
+ * PHP version 5.6, 7
+ *
+ * @author    Elmer Thomas <dx@sendgrid.com>
+ * @copyright 2017 SendGrid
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @version   GIT: <git_id>
+ * @link      http://packagist.org/packages/sendgrid/sendgrid
+ */
+
+namespace Helpers;
+
+class OpenTracking implements \JsonSerializable
+{
+    private $enable;
+    private $substitution_tag;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setSubstitutionTag($substitution_tag)
+    {
+        $this->substitution_tag = $substitution_tag;
+    }
+
+    public function getSubstitutionTag()
+    {
+        return $this->substitution_tag;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable'           => $this->getEnable(),
+                'substitution_tag' => $this->getSubstitutionTag()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
