+++ lib/Helpers/SpamCheck.php
@@ -1,0 +1,65 @@
+<?php
+/**
+ * This helper builds the request body for a /mail/send API call.
+ *
+ * PHP version 5.6, 7
+ *
+ * @author    Elmer Thomas <dx@sendgrid.com>
+ * @copyright 2017 SendGrid
+ * @license   https://opensource.org/licenses/MIT The MIT License
+ * @version   GIT: <git_id>
+ * @link      http://packagist.org/packages/sendgrid/sendgrid
+ */
+
+namespace Helpers;
+
+class SpamCheck implements \JsonSerializable
+{
+    private $enable;
+    private $threshold;
+    private $post_to_url;
+
+    public function setEnable($enable)
+    {
+        $this->enable = $enable;
+    }
+
+    public function getEnable()
+    {
+        return $this->enable;
+    }
+
+    public function setThreshold($threshold)
+    {
+        $this->threshold = $threshold;
+    }
+
+    public function getThreshold()
+    {
+        return $this->threshold;
+    }
+
+    public function setPostToUrl($post_to_url)
+    {
+        $this->post_to_url = $post_to_url;
+    }
+
+    public function getPostToUrl()
+    {
+        return $this->post_to_url;
+    }
+
+    public function jsonSerialize()
+    {
+        return array_filter(
+            [
+                'enable'      => $this->getEnable(),
+                'threshold'   => $this->getThreshold(),
+                'post_to_url' => $this->getPostToUrl()
+            ],
+            function ($value) {
+                return $value !== null;
+            }
+        ) ?: null;
+    }
+}
