+++ lib/loader.php
@@ -3,4 +3,10 @@
  * Allows us to include one file instead of two when working without composer.
  */
 require_once __DIR__ . '/SendGrid.php';
-require_once __DIR__ . '/helpers/mail/Mail.php';
+
+/**
+ * Autoload the helper classes.
+ */
+spl_autoload_register(function ($class) {
+    require_once __DIR__ . '/' . str_replace('\\', '/', $class) . '.php';
+});
